Name change from hyperspy_cl --> lumispy
