+++ README.rst
@@ -1,5 +1,5 @@
-hyperspy_cl is an open-source python library for multi-dimensional analysis of cathodoluminesence data.
+lumispy is an open-source python library for multi-dimensional analysis of cathodoluminesence data.
 
-hyperspy_cl is an extension of the hyperspy package.
+lumispy is an extension of the hyperspy package.
 
-hyperspy_cl is released under the GPL v3 license.
+lumispy is released under the GPL v3 license.
