+++ lumispy/release_info.py
@@ -1,7 +1,7 @@
-name = "hyperspy_cl"
+name = "lumispy"
 version = "0.1.0"
 author = "Duncan Johnstone, Jordi Ferrer Orri"
-copyright = "Copyright 2019, hyperspy_cl Developers"
+copyright = "Copyright 2019, LumiSpy developers"
 credits = ["Duncan Johnstone"]
 license = "GPLv3"
 maintainer = "Duncan Johnstone, Jordi Ferrer Orri"
