+++ setup.py
@@ -1,25 +1,25 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
-# Copyright 2019 The hyperspy_cl developers
+# Copyright 2019 The LumiSpy developers
 #
-# This file is part of hyperspy_cl.
+# This file is part of lumispy.
 #
-# hyperspy_cl is free software: you can redistribute it and/or modify
+# lumispy is free software: you can redistribute it and/or modify
 # it under the terms of the GNU General Public License as published by
 # the Free Software Foundation, either version 3 of the License, or
 # (at your option) any later version.
 #
-# hyperspy_cl is distributed in the hope that it will be useful,
+# lumispy is distributed in the hope that it will be useful,
 # but WITHOUT ANY WARRANTY; without even the implied warranty of
 # MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 # GNU General Public License for more details.
 #
 # You should have received a copy of the GNU General Public License
-# along with hyperspy_cl.  If not, see <http://www.gnu.org/licenses/>.
+# along with lumispy.  If not, see <http://www.gnu.org/licenses/>.
 
 from setuptools import setup, find_packages
 
-exec(open('hyperspy_cl/release_info.py').read())  # grab version info
+exec(open('lumispy/release_info.py').read())  # grab version info
 
 
 setup(
@@ -29,7 +29,7 @@ setup(
     author=author,
     author_email=email,
     license=license,
-    url="https://github.com/hyperspy_cl/hyperspy_cl",
+    url="https://github.com/lumispy/lumispy",
     long_description=open('README.rst').read(),
     classifiers=[
         "Programming Language :: Python :: 3",
@@ -54,7 +54,7 @@ setup(
     ],
     package_data={
         "": ["LICENSE", "README.rst"],
-        "hyperspy_cl": ["*.py", "hyperspy_extension.yaml"],
+        "lumispy": ["*.py", "hyperspy_extension.yaml"],
     },
-    entry_points={'hyperspy.extensions': ['hyperspy_cl = hyperspy_cl']},
+    entry_points={'hyperspy.extensions': ['lumispy = lumispy']},
     )
